+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandListener.java
@@ -2,9 +2,13 @@ package com.github.johnnyjayjay.discord.commandapi;
 
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
+import net.dv8tion.jda.core.entities.Member;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static java.lang.String.join;
 import static java.lang.String.format;
@@ -13,9 +17,11 @@ import static java.lang.String.format;
 class CommandListener extends ListenerAdapter {
 
     private CommandSettings settings;
+    private Map<Long, Long> cooldowns; // Long: User id, Long: last timestamp
 
     public CommandListener(CommandSettings settings) {
         this.settings = settings;
+        this.cooldowns = new HashMap<>();
     }
 
     @Override
@@ -24,18 +30,23 @@ class CommandListener extends ListenerAdapter {
         String prefix = settings.getPrefix(event.getGuild().getIdLong());
         if (!event.getAuthor().isBot()) {
             if (raw.startsWith(prefix)) {
+                long timestamp = System.currentTimeMillis();
+                long userId = event.getAuthor().getIdLong();
+                if (cooldowns.containsKey(userId) && (timestamp - cooldowns.get(userId)) < settings.getCooldown())
+                    return;
+                cooldowns.put(userId, timestamp);
                 CommandEvent.Command cmd = new CommandEvent.Command(raw, prefix, settings);
                 if (cmd.getExecutor() != null) {
                     cmd.getExecutor().onCommand(new CommandEvent(event.getJDA(), event.getResponseNumber(), event.getMessage(), cmd),
                             event.getMember(), event.getChannel(), cmd.getArgs());
                 } else if (settings.getHelpLabels().contains(cmd.getLabel()) && event.getGuild().getSelfMember().hasPermission(event.getChannel(), Permission.MESSAGE_WRITE)) {
-                    this.sendInfo(event.getChannel(), prefix, cmd.getArgs());
+                    this.sendInfo(event.getMember(), event.getChannel(), prefix, cmd.getArgs());
                 }
             }
         }
     }
 
-    private void sendInfo(TextChannel channel, String prefix, String[] args) {
+    private void sendInfo(Member member, TextChannel channel, String prefix, String[] args) {
         var builder = new EmbedBuilder().setTitle("Help");
         if (args.length == 0) {
             String helpLabels = format("[%s]", join("|", settings.getHelpLabels().toArray(new String[settings.getHelpLabels().size()])));
@@ -46,7 +57,7 @@ class CommandListener extends ListenerAdapter {
             channel.sendMessage(builder.build()).queue();
         } else if (args.length == 1 && settings.getCommands().containsKey(args[0])) {
             builder.appendDescription(format("**Command Info for:** `%s`\n\n", args[0]))
-                    .appendDescription(settings.getCommands().get(args[0]).info(channel.getGuild()));
+                    .appendDescription(settings.getCommands().get(args[0]).info(member));
             channel.sendMessage(builder.build()).queue();
         }
     }